requirements.txt
Django==4.2.7
environs==9.5.0
pyTelegramBotAPI==4.14.0
django-phonenumber-field==7.2.0
phonenumberslite==8.13.23
